% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_history.R
\name{collect_cohorts}
\alias{collect_cohorts}
\title{Function that collects cohorts for analysis}
\usage{
collect_cohorts(execution_settings, targetCohortId)
}
\arguments{
\item{execution_settings}{executionSettings object with connection credentials}

\item{targetCohortId}{an integer specifying the target cohort}
}
\description{
pulls down entire cohort table for person in the target cohort
}
