% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTreatmentPathways.R
\name{computeTreatmentPathways}
\alias{computeTreatmentPathways}
\title{Compute treatment pathways from a set of event cohorts}
\usage{
computeTreatmentPathways(
  cdm,
  targetCohortTable,
  eventCohortTable,
  periodPriorToIndex = 0,
  minEraDuration = 0,
  eraCollapseSize = 1,
  verbose = TRUE,
  maxIterations = 1000
)
}
\arguments{
\item{cdm}{A cdm reference object with at least two cohort tables: A target
cohort table and an event cohort table.}

\item{targetCohortTable}{The name of the target cohort table}

\item{eventCohortTable}{The name of the event cohort table}

\item{periodPriorToIndex}{The number of days prior to index to look for event
cohort start dates.}

\item{minEraDuration}{Minimum time an event era or combination}

\item{eraCollapseSize}{Maximum time interval between treatments
should last to be included in analysis}

\item{verbose}{Should progress messages be printed?}

\item{maxIterations}{Maximum number of iterations to allow for collapse algorithm}
}
\value{
A cdm reference with a new table treatment_patterns added
}
\description{
Compute treatment pathways from a set of event cohorts
}
