% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_settings.R, R/stratafication.R
\name{create_concept_stratification}
\alias{create_concept_stratification}
\title{A function to create a concept strata}
\usage{
create_concept_stratification(name, domain, concept)

create_concept_stratification(name, domain, concept)
}
\arguments{
\item{name}{the name of the strata}

\item{domain}{the domain of the concept}

\item{concept}{the OMOP concept id for the strata}
}
\value{
a concept strata object

a concept strata object
}
\description{
A function to create a concept strata

A function to create a concept strata
}
