% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_settings.R, R/stratafication.R
\name{create_date_stratification}
\alias{create_date_stratification}
\title{A function to create a date strata}
\usage{
create_date_stratification(name, op, value)

create_date_stratification(name, op, value)
}
\arguments{
\item{name}{the name of the strata}

\item{op}{the operator for the strata}

\item{value}{the value for the date strata}
}
\value{
a date strata object

a date strata object
}
\description{
A function to create a date strata

A function to create a date strata
}
