% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeToEventAnalysis.R
\name{create_survival_table}
\alias{create_survival_table}
\title{Function that creates survival tables from the treatment history}
\usage{
create_survival_table(
  treatment_history,
  connectionDetails,
  resultsDatabaseSchema,
  cohortTable,
  analysisSettings
)
}
\arguments{
\item{treatment_history}{the dataframe output from create_treatment_history}

\item{connectionDetails}{a list of connectionDetails for DatabaseConnector}

\item{resultsDatabaseSchema}{the schema that hosts the users writeable results tables (or scratch), defaults to value
in config.yml of the active configuration}

\item{cohortTable}{the name where the cohorts are stored in the results schema}

\item{analysisSettings}{a DrugUtilizationAnalysisSettings object that defines the elements of the analysis}
}
\value{
a SurvivalAnalysis object containing the data for the survival
analysis, the survfit object and the surv_summary object from survminer
}
\description{
Function that creates survival tables from the treatment history
}
