% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTreatmentHistory.R
\name{create_treatment_history}
\alias{create_treatment_history}
\title{Function that creates the treatment history}
\usage{
create_treatment_history(
  connectionDetails,
  cdmDatabaseSchema,
  resultsDatabaseSchema,
  database,
  cohortTable,
  analysisSettings
)
}
\arguments{
\item{connectionDetails}{a list of connectionDetails for DatabaseConnector}

\item{cdmDatabaseSchema}{the schema that hosts the cdm tables, defaults to value
in config.yml of the active configuration}

\item{resultsDatabaseSchema}{the schema that hosts the users writeable results tables (or scratch), defaults to value
in config.yml of the active configuration}

\item{cohortTable}{the name where the cohorts are stored in the results schema}

\item{analysisSettings}{a DrugUtilizationAnalysisSettings object that defines the elements of the analysis}
}
\value{
a data.table with the treatment history
}
\description{
This function adapts the underlying calculations for treatment patterns from TreatmentPatterns package
}
