% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayAnalysis.R
\name{create_treatment_patterns}
\alias{create_treatment_patterns}
\title{Function to construct the treatment patterns}
\usage{
create_treatment_patterns(treatment_history, analysisSettings)
}
\arguments{
\item{treatment_history}{the dataframe output from create_treatment_history}

\item{analysisSettings}{a DrugUtilizationAnalysisSettings object that defines the elements of the analysis}
}
\value{
a PathwayAnalysis object with the treatment patterns and attrition
}
\description{
Function to construct the treatment patterns
}
