% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTreatmentPathways.R
\name{createTestCdm}
\alias{createTestCdm}
\title{Create an example cdm with event cohorts}
\usage{
createTestCdm(n = NULL)
}
\arguments{
\item{n}{Number of unique persons to include in the target and event cohorts
By default all persons with a GI bleed condition will be included.}
}
\value{
A cdm reference object with two cohort tables: target_cohort and event_cohorts.
}
\description{
Return an example cdm object with a target cohort of patients with GIBleed
and event cohorts for every drug ingredient exposure in the Eunomia test
database.
}
\examples{
\dontrun{
cdm <- createTestCdm()
cdm <- createTestCdm(n = 10)
}
}
