% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{defineTreatmentHistory}
\alias{defineTreatmentHistory}
\title{Define the treatment history settings}
\usage{
defineTreatmentHistory(cohorts, ...)
}
\arguments{
\item{cohorts}{a tibble of cohorts that are built manually or using the identifyThCohorts
function from this package. A valid cohorts tibble has a cohort_id, cohort_name and type columns.}

\item{...}{a dot list of other options to modify in the treatment history settings}
}
\description{
This function defines the parameters for an individual treatment history analysis.
These settings are based on TreatmentPatterns package available from DARWIN.
Upon cohort specification, the user can input parameters for the treatment history construction.
A treatment history data.frame is constructed based on the following options:
\itemize{
\item{minEraDuration}{the minimum number of days for an event era}
\item{eraCollapseSize}{the maximum gap, defined in days, between two eras of the same event cohort which would still allow the eras to be collapesed into one era}
\item{combinationWindow}{the maximum gap, defined in days, between two eras of the same event cohort which would still allow the eras to be collapesed into one era}
\item{minPostCombinationDuration}{the minimum time, defined in days, that an event era before or after a generated combination treatment should last to be included in the pathway as a separate treatment}
\item{filterTreatments}{a character string describing the filtering preference for including into a pathway.}
}
Options for filterTreatments include 'First" which limits the pathway to the first occurrence of a treatment,
'Changes' which removed sequential repeated treatments or 'All' which includes all treatments.
There are additional options such as maxPathLength, minCellCount, minCellMethod, groupCombinations,
periodPriorToIndex, and addNoPaths which are trivial but can be modified.
}
