% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugUtilizationMod-package.R
\docType{package}
\name{drugUtilizationMod-package}
\alias{drugUtilizationMod}
\alias{drugUtilizationMod-package}
\title{drugUtilizationMod: Module for Drug Utilization Studies in OHDSI}
\description{
An R package that runs the necessary components of an OHDSI drug utilization study
}
\author{
\strong{Maintainer}: Martin Lavallee \email{martin.lavallee@odysseusinc.com}

}
\keyword{internal}
