% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{findPatterns}
\alias{findPatterns}
\title{Function to determine treatment patterns}
\usage{
findPatterns(th, minNumPatterns = 30)
}
\arguments{
\item{th}{the treatment history data.frame}

\item{minNumPatterns}{the minimum number of patterns needed to remain as a eligible pathway}
}
\description{
this function aggregates the treatment history into a
treatment patterns analysis. The function returns information on the attrition,
a data structure for a sankey diagram and a clean table of the treatment patterns.

The treatment patterns output can be saved or used to create a sankey diagram
}
