% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{identifyThCohorts}
\alias{identifyThCohorts}
\title{Identify Treatment history cohorts}
\usage{
identifyThCohorts(cohortMeta, targetId, eventIds)
}
\arguments{
\item{cohortMeta}{meta table describing the cohorts in the cohort table}

\item{targetId}{a cohort id integer identifying the target cohort for the
treatment history analysis to search in the cohortMeta}

\item{eventIds}{an integer vector of cohortIds used to look identifying the
events cohorts for the treatment history analysis to searach in the cohortMeta}
}
\description{
this function takes the meta file for all cohorts in the cohort table
and allows the user to identify which is the target cohort and which are the
event cohorts
}
