% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeToEventAnalysis.R
\name{plot_kaplan_meier}
\alias{plot_kaplan_meier}
\title{Function to plot the survival analysis for time to discontinuation}
\usage{
plot_kaplan_meier(
  survival_table,
  type = c("Single", "Combinations", "All"),
  selectTop = 7
)
}
\arguments{
\item{survival_table}{a SurvivalAnalysis object with information about the survival analysis}

\item{type}{identify the treatment line type to display, Either single, only combinations or all lines}

\item{selectTop}{display an integer number of lines of treatment to display in the kaplan meier plot}
}
\value{
a kaplan meier plot suing survminer. We implement the lancet color schema from ggsci
}
\description{
Function to plot the survival analysis for time to discontinuation
}
