% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{plot_patterns}
\alias{plot_patterns}
\title{Function to plot sankey}
\usage{
plot_patterns(treatmentPatterns)
}
\arguments{
\item{treatmentPatterns}{object to use for sankey}
}
\description{
a sankey diagram describing the treatment patterns
}
