% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayAnalysis.R
\name{plot_treatment_patterns}
\alias{plot_treatment_patterns}
\title{Function to plot the sankey for treatment patterns}
\usage{
plot_treatment_patterns(pathway_analysis, type = "sankey")
}
\arguments{
\item{pathway_analysis}{a PathwayAnalysis object with information about the treatment patterns}
}
\value{
a sankey diagram using networkD3::sankeyNetwork. We implement the lancet color schema from ggsci
}
\description{
Function to plot the sankey for treatment patterns
}
