% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeToEventAnalysis.R
\name{save_survival_table}
\alias{save_survival_table}
\title{Function to save the survival analysis as multiple csv}
\usage{
save_survival_table(survival_table, analysisSettings)
}
\arguments{
\item{survival_table}{a SurvivalAnalysis object with information about the survival analysis}

\item{analysisSettings}{a DrugUtilizationAnalysisSettings object that defines the elements of the analysis}
}
\description{
Function to save the survival analysis as multiple csv
}
