% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTreatmentHistory.R
\name{save_treatment_history}
\alias{save_treatment_history}
\title{Function to save the treatment history as csv}
\usage{
save_treatment_history(treatment_history, analysisSettings)
}
\arguments{
\item{treatment_history}{the dataframe output from create_treatment_history}

\item{analysisSettings}{a DrugUtilizationAnalysisSettings object that defines the elements of the analysis}
}
\description{
Function to save the treatment history as csv
}
