% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayAnalysis.R
\name{save_treatment_patterns}
\alias{save_treatment_patterns}
\title{Function to save the treatment patterns as multiple csv}
\usage{
save_treatment_patterns(treatment_patterns, analysisSettings)
}
\arguments{
\item{analysisSettings}{a DrugUtilizationAnalysisSettings object that defines the elements of the analysis}

\item{pathway_analysis}{a PathwayAnalysis object with information about the treatment patterns}
}
\description{
Function to save the treatment patterns as multiple csv
}
