% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_settings.R, R/stratafication.R
\name{stratification_list}
\alias{stratification_list}
\title{#' A function to create a cohort strata
#'
#' @param name the name of the strata
#' @param concept the OMOP concept id for the strata
#' @return a concept strata object
#' @export
create_cohort_stratification <- function(name,
cohort) {
structure(list(name = name, cohort = cohort), class = "cohortStrata")
}
A function to create strata list object}
\usage{
stratification_list(...)

stratification_list(...)
}
\arguments{
\item{...}{a list of strata objects}
}
\value{
a strata list object

a strata list object
}
\description{
#' A function to create a cohort strata
#'
#' @param name the name of the strata
#' @param concept the OMOP concept id for the strata
#' @return a concept strata object
#' @export
create_cohort_stratification <- function(name,
cohort) {
structure(list(name = name, cohort = cohort), class = "cohortStrata")
}
A function to create strata list object

#' A function to create a cohort strata
#'
#' @param name the name of the strata
#' @param concept the OMOP concept id for the strata
#' @return a concept strata object
#' @export
create_cohort_stratification <- function(name,
cohort) {
structure(list(name = name, cohort = cohort), class = "cohortStrata")
}
A function to create strata list object
}
