% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_history.R
\name{th}
\alias{th}
\title{Function that constructs the treatment history}
\usage{
th(execution_settings, treatmentHistorySettings)
}
\arguments{
\item{execution_settings}{executionSettings object with connection credentials}

\item{treatmentHistorySettings}{a object with the treatment history settings}
}
\description{
this function creates the treatment history data.frame
based on the input of the treamtent history settings
}
